% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{quant_pal}
\alias{quant_pal}
\title{Creates custom color palettes}
\usage{
quant_pal(palette = c("main", "withgrey", "pnl", "weights", "metro",
  "greyscale", "heatmap"), reverse = FALSE, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Creates custom color palettes
}
