% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_quant}
\alias{scale_color_quant}
\title{Sets colors in ggplot2 plots}
\usage{
scale_color_quant(palette = c("main", "withgrey", "pnl", "weights", "metro",
  "greyscale", "heatmap"), discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
Sets colors in ggplot2 plots
}
\details{
Four color palettes are defined. By default discrete colors are chosen from the palette, otherwise a color gradient is generated
}
