% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_quant_manual}
\alias{scale_color_quant_manual}
\title{Set manual color values}
\usage{
scale_color_quant_manual(palette = c("main", "withgrey", "pnl", "weights",
  "metro", "greyscale", "heatmap"), values = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Set manual color values
}
\details{
Four color palettes are defined. If values parameter is not set, it is determined from the palette parameter
}
