% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{test_pal}
\alias{test_pal}
\title{Test a color palette using a pie chart}
\usage{
test_pal(palette = c("main", "withgrey", "pnl", "weights", "metro",
  "greyscale", "heatmap"), reverse = FALSE, n = 5)
}
\arguments{
\item{n}{}
}
\description{
Test a color palette using a pie chart
}
