% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_quant.R
\name{theme_quant}
\alias{theme_quant}
\title{Defines a ggplot2 custom theme}
\usage{
theme_quant(base_size = 11, strip_text_size = 12, strip_text_margin = 5,
  subtitle_size = 12, subtitle_margin = 10, plot_title_size = 18,
  plot_title_margin = 10, ...)
}
\arguments{
\item{...}{}
}
\description{
Defines a ggplot2 custom theme
}
